move header to Dashboard, fix iOS layout bug
